% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{BMD_continuous_MA-class}
\alias{BMD_continuous_MA-class}
\alias{BMD_continuous_MA}
\title{Model averaged continuous fit class}
\value{
A BMD_continuous_MA object
}
\description{
Methods include plot, MAdensity_plot, and cleveland_plot
}
\section{Slots}{

\describe{
\item{\code{models}}{List of BMD_continuous_fit_* objects (one per model)}

\item{\code{bmd}}{Named vector of model averaged BMD, BMDL, BMDU}

\item{\code{ma_bmd}}{Matrix of spline BMD CDF}

\item{\code{posterior_probs}}{Named vector of posterior probabilities of each model}

\item{\code{type}}{Character string of fit_type}
}}

\examples{
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    normprior(0, 10, -100, 100)
)
p_hill_norm <- create_continuous_prior(prior, "hill", "normal")
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    lnormprior(0, 1, 0, 100),
    normprior(0, 10, -100, 100)
)
p_hill_ncv <- create_continuous_prior(prior, "hill", "normal-ncv")
prior_list <- list(p_hill_norm, p_hill_ncv)
doses <- cbind(c(0, 25, 50, 100, 200))
y <- cbind(
    c(6, 5.2, 2.4, 1.1, 0.75),
    c(20, 20, 19, 20, 20),
    c(1, 1.1, 0.81, 0.74, 1)
)
model <- ma_continuous_fit(doses, y,
    fit_type = "mcmc",
    BMR_TYPE = "abs",
    BMR = 0.1,
    samples = 1e4,
    model_list = prior_list
)
model
}
