% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\docType{class}
\name{BMD_dichotomous_fit_MA-class}
\alias{BMD_dichotomous_fit_MA-class}
\alias{BMD_dichotomous_fit_MA}
\title{BMD Bayesian Dichotomous Model class}
\value{
A BMD_dichotomous_fit_MA object
}
\description{
Methods include plot, cleveland_plot
}
\section{Slots}{

\describe{
\item{\code{bmd}}{Named vector of model averaged BMD, BMDL, BMDU}

\item{\code{posterior_probs}}{Named vector of posterior probabilities of each model}

\item{\code{submodels}}{list of all of models used in model averaging}

\item{\code{fit_type}}{string value specifying fit type "mcmc" or "laplace"}

\item{\code{full_model}}{Full model name}
}}

\examples{
mData <- matrix(c(
    0, 2, 50,
    1, 2, 50,
    3, 10, 50,
    16, 18, 50,
    32, 18, 50,
    33, 17, 50
), nrow = 6, ncol = 3, byrow = TRUE)
D <- mData[, 1]
Y <- mData[, 2]
N <- mData[, 3]
model <- ma_dichotomous_fit(D, Y, N)

model
}
