% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{BMD_dichotomous_MA-class}
\alias{BMD_dichotomous_MA-class}
\title{Model averaged dichotomous fit class}
\value{
A BMD_dichotomous_MA object
}
\description{
Methods include plot, MAdensity_plot, and cleveland_plot
}
\section{Slots}{

\describe{
\item{\code{models}}{List of BMD_dichotomous_fit_* objects (one per model)}

\item{\code{bmd}}{Named vector of model averaged BMD, BMDL, BMDU}

\item{\code{ma_bmd}}{Matrix of spline BMD CDF}

\item{\code{posterior_probs}}{Named vector of posterior probabilities of each model}

\item{\code{type}}{Character string of fit_type}

\item{\code{BMD_CDF}}{Matrix of spline BMD CDF}
}}

