% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cleveland_plot}
\alias{cleveland_plot}
\alias{cleveland_plot,BMD_dichotomous_MA-method}
\alias{cleveland_plot,BMD_continuous_MA-method}
\title{cleveland_plot - Create a Cleveland plot from a model averaged model.}
\usage{
cleveland_plot(A)

\S4method{cleveland_plot}{BMD_dichotomous_MA}(A)

\S4method{cleveland_plot}{BMD_continuous_MA}(A)
}
\arguments{
\item{A}{the model averaged model to plot. An \code{\link{BMD_continuous_MA}}
or \code{\link{BMD_dichotomous_MA}} object.}
}
\value{
Returns a \code{ggplot2} graphics object.
}
\description{
Create a Cleveland plot from a model averaged model.
}
\examples{
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    normprior(0, 10, -100, 100)
)
p_hill_norm <- create_continuous_prior(prior, "hill", "normal")
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    lnormprior(0, 1, 0, 100),
    normprior(0, 10, -100, 100)
)
p_hill_ncv <- create_continuous_prior(prior, "hill", "normal-ncv")
prior_list <- list(p_hill_norm, p_hill_ncv)
doses <- cbind(c(0, 25, 50, 100, 200))
y <- cbind(
    c(6, 5.2, 2.4, 1.1, 0.75),
    c(20, 20, 19, 20, 20),
    c(1, 1.1, 0.81, 0.74, 1)
)
model <- ma_continuous_fit(doses, y,
    fit_type = "mcmc",
    BMR_TYPE = "abs",
    BMR = 0.1,
    samples = 1e4,
    model_list = prior_list
)
cleveland_plot(model)
}
