% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{get_prior}
\alias{get_prior}
\title{Gets the prior of a BMD_continuous_fit}
\usage{
get_prior(object)
}
\arguments{
\item{oject}{A single continuous fit object}
}
\value{
Returns the included priorClass
}
\description{
This is equivalent to object@prior, but the accessor is
needed to avoid BiocCheck warnings for the vignette.
}
\examples{
M2 <- matrix(0, nrow = 5, ncol = 4)
colnames(M2) <- c("Dose", "Resp", "N", "StDev")
M2[, 1] <- c(0, 25, 50, 100, 200)
M2[, 2] <- c(6, 5.2, 2.4, 1.1, 0.75)
M2[, 3] <- c(20, 20, 19, 20, 20)
M2[, 4] <- c(1.2, 1.1, 0.81, 0.74, 0.66)
model <- single_continuous_fit(
    M2[, 1, drop = FALSE],
    M2[, 2:4],
    BMR_TYPE = "sd",
    BMR = 1, ewald = TRUE,
    distribution = "normal",
    fit_type = "laplace",
    model_type = "hill",
    threads = 2,
    seed = 12331
)

get_prior(model)
}
