% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{class}
\name{NtpDunn-class}
\alias{NtpDunn-class}
\alias{NtpDunn}
\title{NTP Dunn class}
\value{
A NtpDunn object
}
\description{
NTP Dunn class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame containing results of the NTP Dunn test}
}}

