% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{class}
\name{NtpDunnett-class}
\alias{NtpDunnett-class}
\alias{NtpDunnett}
\title{NTP Dunnett class}
\value{
A NtpDunnett object
}
\description{
NTP Dunnett class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame containing results of the NTP Dunnett test}
}}

