% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{class}
\name{NtpPolyk-class}
\alias{NtpPolyk-class}
\alias{NtpPolyk}
\title{NTP PolyK class}
\value{
A NtpPolyk object
}
\description{
NTP PolyK class
}
\section{Slots}{

\describe{
\item{\code{dose}}{An equation of the form \eqn{Y \sim X.} Here the variable
\eqn{Y} is the response of interest, and \eqn{X}
represents discrete experimental
conditions. For example, if weight is the dependent variable, and you are
interested in looking at the trend across sex one would have 'weight ~ sex'.}

\item{\code{tumor}}{A vector with column names in the formula.}

\item{\code{daysOnStudy}}{The name of the variable containing the doses in the
data frame \eqn{data}. It is expected multiple doses for each of the
experimental conditions \eqn{X}.}

\item{\code{results}}{A matrix with results of a PolyK trend test for each level
 in \eqn{dose\_name}.
More information on this procedure at: \doi{10.2307/2531856} and
\doi{10.2307/2532200}
This procedure returns a vector of three p-values for the poly-1.5,
 poly-3, and poly-6 test
respectively.}
}}

