% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{class}
\name{NtpShirley-class}
\alias{NtpShirley-class}
\alias{NtpShirley}
\title{NTP Shirley class}
\value{
A NtpShirley object
}
\description{
NTP Shirley class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame containing results of the NTP Shirley test}
}}

