% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{class}
\name{NtpWilliams-class}
\alias{NtpWilliams-class}
\alias{NtpWilliams}
\title{NTP Williams class}
\value{
A NtpWilliams object
}
\description{
NTP Williams class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame containing results of the NTP Williams test}
}}

