% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R, R/models.R
\name{plot,BMD_continuous_fit_MCMC,missing-method}
\alias{plot,BMD_continuous_fit_MCMC,missing-method}
\alias{plot,BMD_continuous_fit_maximized,missing-method}
\alias{plot,BMD_dichotomous_fit_MCMC,missing-method}
\alias{plot,BMD_dichotomous_fit_maximized,missing-method}
\alias{plot,BMD_continuous_MA,missing-method}
\alias{plot,BMD_dichotomous_MA,missing-method}
\title{Plot method for BMD_continuous_fit_MCMC objects}
\usage{
\S4method{plot}{BMD_continuous_fit_MCMC,missing}(x, y, ...)

\S4method{plot}{BMD_continuous_fit_maximized,missing}(x, y, ...)

\S4method{plot}{BMD_dichotomous_fit_MCMC,missing}(x, y, ...)

\S4method{plot}{BMD_dichotomous_fit_maximized,missing}(x, y, ...)

\S4method{plot}{BMD_continuous_MA,missing}(x, y, ...)

\S4method{plot}{BMD_dichotomous_MA,missing}(x, y, ...)
}
\arguments{
\item{x}{S4 object BMD_dichotomous_MA}

\item{y}{NA / missing}

\item{...}{additional parameters (not used)}
}
\value{
ggplot object

ggplot object

ggplot object
}
\description{
Plot method for BMD_continuous_fit_MCMC objects

Plot method for BMD_continuous_fit_maximized objects

Plot method for BMD_dichotomous_fit_MCMC objects

Plot method for BMD_dichotomous_fit_maximized objects
}
