% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ToxicRToo}
\alias{predict.ToxicRToo}
\title{Predict Class Helper Function}
\usage{
predict.ToxicRToo(x, new_doses = NULL, ...)
}
\arguments{
\item{x}{Single continuous/dichotomous model fit object}

\item{new_doses}{New doses to predict the response at}

\item{...}{Other unused arguments}
}
\value{
Predicted response values for the model
}
\description{
While simply using predict() should work on single fits, this 
function provides the documentation for specifying new data with
a named parameter called new_doses.
}
