% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{priorClass-class}
\alias{priorClass-class}
\alias{priorClass}
\title{The prior class defining a prior for a specific model}
\value{
A priorClass object
}
\description{
The prior class defining a prior for a specific model
}
\section{Slots}{

\describe{
\item{\code{prior}}{A matrix of prior information
(see \code{\link{create_prior_list}})}

\item{\code{model}}{Character string of model description}

\item{\code{mean}}{Character string of model type (e.g. hill)}

\item{\code{parameters}}{Character vector of parameter names}

\item{\code{distribution}}{Character string of the error distribution}
}}

\examples{
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    normprior(0, 10, -100, 100)
)
p_hill_norm <- create_continuous_prior(prior, "hill", "normal")
prior <- create_prior_list(
    normprior(0, 1, -100, 100),
    normprior(0, 1, -1e4, 1e4),
    lnormprior(0, 1, 0, 100),
    lnormprior(log(1), 0.4215, 0, 18),
    lnormprior(0, 1, 0, 100),
    normprior(0, 10, -100, 100)
)
p_hill_ncv <- create_continuous_prior(prior, "hill", "normal-ncv")
prior_list <- list(p_hill_norm, p_hill_ncv)
doses <- cbind(c(0, 25, 50, 100, 200))
y <- cbind(
    c(6, 5.2, 2.4, 1.1, 0.75),
    c(20, 20, 19, 20, 20),
    c(1, 1.1, 0.81, 0.74, 1)
)
model <- ma_continuous_fit(doses, y,
    fit_type = "mcmc",
    BMR_TYPE = "abs",
    BMR = 0.1,
    samples = 1e4,
    model_list = prior_list
)
model
}
