% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\docType{class}
\name{BMD_continuous_fit_MCMC-class}
\alias{BMD_continuous_fit_MCMC-class}
\alias{BMD_continuous_fit_MCMC}
\title{Single continuous model MCMC fit class}
\value{
A BMD continuous fit (MCMC) object
}
\description{
Methods include plot and predict
}
\section{Slots}{

\describe{
\item{\code{bmd}}{A vector containing the benchmark dose (BMD) and
\eqn{100\times(1-2\alpha)} confidence intervals.}

\item{\code{data}}{Matrix of input data}

\item{\code{prior}}{Prior information}

\item{\code{model}}{String specifying the mean model used.}

\item{\code{options}}{Numeric vector of options used in the fitting procedure.}

\item{\code{mcmc_result}}{List of 2 matrices: BMD and parameter samples}

\item{\code{covariance}}{The variance-covariance matrix for the parameters.}

\item{\code{maximum}}{Maximum value of likelihood/posterior}

\item{\code{bmd_dist}}{Quantiles for the BMD distribution.}

\item{\code{fitted_model}}{A list with a simplified S3 object}

\item{\code{transformed}}{Are the data \eqn{\log(x+\sqrt{x^2+1})} transformed?}

\item{\code{full_model}}{The model along with the likelihood distribution.}

\item{\code{parameters}}{The parameter estimates produced by the procedure, which
are relative to the model given in \code{full_model}.  The last parameter
is always the estimate for \eqn{\log(sigma^2)}.}
}}

\examples{
M2 <- matrix(0, nrow = 5, ncol = 4)
colnames(M2) <- c("Dose", "Resp", "N", "StDev")
M2[, 1] <- c(0, 25, 50, 100, 200)
M2[, 2] <- c(6, 5.2, 2.4, 1.1, 0.75)
M2[, 3] <- c(20, 20, 19, 20, 20)
M2[, 4] <- c(1.2, 1.1, 0.81, 0.74, 0.66)
model <- single_continuous_fit(
    M2[, 1, drop = FALSE],
    M2[, 2:4],
    BMR_TYPE = "sd",
    BMR = 1, ewald = TRUE,
    distribution = "normal",
    fit_type = "mcmc",
    model_type = "hill",
    threads = 2,
    seed = 12331
)

model
}
