% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{priorClass-class}
\alias{priorClass-class}
\title{The prior class defining a prior for a specific model}
\value{
A priorClass object
}
\description{
The prior class defining a prior for a specific model
}
\section{Slots}{

\describe{
\item{\code{prior}}{A matrix of prior information (see \code{\link{create_prior_list}})}

\item{\code{model}}{Character string of model description}

\item{\code{mean}}{Character string of model type (e.g. hill)}

\item{\code{parameters}}{Character vector of parameter names}

\item{\code{distribution}}{Character string of the error distribution}
}}

